Labels.py changes

+++ labels.py
@@ -6,6 +6,10 @@ import cv2
 import math
 import os
 import shutil
+import random
+import pybboxes as pbx
+
+
 
 def yolobbox2bbox(x,y,w,h):
     x1, y1 = (x-(w/2))*640, (y-(h/2))*640
@@ -30,8 +34,11 @@ def create_dir(path):
         os.makedirs(path)
 
 # Specify the folder path
-folder_path = filedialog.askdirectory(title="Select the image folder",
+folder_path_img = filedialog.askdirectory(title="Select the IMAGE folder",
                                                     initialdir=os.path.dirname(os.path.abspath(__file__)))
+# Specify the folder path
+folder_path_lbl = filedialog.askdirectory(title="Select the LABEL folder",
+                                                    initialdir=folder_path_img)
 
 #display message box and ask yes or no
 answer = messagebox.askyesno("Confirmation", "Do you want to visually control the cropping process? If yes, each image will be displayed for a short while so you can make sure there are no discrepancies. If no, the process will be faster.")
@@ -40,19 +47,34 @@ if answer:
     frames = 300
 else:
     frames = 1
+
+#display message box and ask yes or no
+answer = messagebox.askyesno("Confirmation", "Do you want to specify separate output folder? If yes, you will be asked to select the folder. If no, the output folder will be the same as the input folder.")
+
+if answer:
+    out_folder_path_img = filedialog.askdirectory(title="Select the IMAGE output folder",
+                                          initialdir=folder_path_img)
+    out_folder_path_lbl = filedialog.askdirectory(title="Select the LABEL output folder",
+                                                  initialdir=folder_path_lbl)
+else:
+    out_folder_path_img = folder_path_img
+    out_folder_path_lbl = folder_path_lbl
+
+
+
 # Specify the paths of the original and destination folders
 tmp_folder = "resources/tmp"
 create_dir(tmp_folder)
 
 # Iterate through all files in the folder
-for filename in os.listdir(folder_path):
+for filename in os.listdir(folder_path_img):
     # Check if the file is an image file (you can customize this condition as needed)
     if filename.endswith('.jpg') or filename.endswith('.png'):
         # Get the full file path
-        file_path = os.path.join(folder_path, filename)
+        file_path = os.path.join(folder_path_img, filename)
 
         # Construct the full paths to the original and destination files
-        original_path = os.path.join(folder_path, filename)
+        original_path = os.path.join(folder_path_img, filename)
         destination_path = os.path.join(tmp_folder, filename)
 
         # Move the file to the destination folder
@@ -62,29 +84,43 @@ for filename in os.listdir(folder_path):
         img = Image.open(destination_path)
 
         # Open the corresponding txt file
-        txt_path = f"{folder_path}/{filename[:-4]}.txt"
+        txt_path = f"{folder_path_lbl}/{filename[:-4]}.txt"
         if not os.path.exists(txt_path):
-            continue
-        with open(txt_path, "r") as f:
-            coords = f.readline().strip().split()[1:]
+            # Calculate the centre of the random point
+            x = random.randint(0, img.size[0])
+            y = random.randint(0, img.size[1])
 
-        # Convert the coordinates to floats and scale them for a 320x320 image
-        coords = [float(coord) for coord in coords]
-        box_left, box_top, box_right, box_bottom = yolobbox2bbox(*coords)
-        box_width = box_right - box_left
-        box_height = box_bottom - box_top
+            # Calculate the coordinates of the cropped image
+            factor = 320
+            x1 = int(max(0, min((x - factor // 2), int(img.size[0]) - factor)))
+            y1 = int(max(0, min((y - factor // 2), int(img.size[1]) - factor)))
+            x2 = int(max(factor, min((x + factor // 2), int(img.size[0]))))
+            y2 = int(max(factor, min((y + factor // 2), int(img.size[1]))))
 
-        # Calculate the centre of the bounding box in the original image
-        x = box_right - box_width/2
-        y = box_bottom - box_height/2
+            box_left, box_top, box_right, box_bottom = x1, y1, x2, y2
+            box_width = box_right - box_left
+            box_height = box_bottom - box_top
+        else:
+            with open(txt_path, "r") as f:
+                coords = f.readline().strip().split()[1:]
 
-        # Calculate the coordinates of the cropped image
-        factor = max(box_width, box_height)
-        factor = max(factor, 320)
-        x1 = int(max(0, min((x - factor // 2), int(img.size[0]) - factor)))
-        y1 = int(max(0, min((y - factor // 2), int(img.size[1]) - factor)))
-        x2 = int(max(factor, min((x + factor // 2), int(img.size[0]))))
-        y2 = int(max(factor, min((y + factor // 2), int(img.size[1]))))
+            # Convert the coordinates to floats and scale them for a 320x320 image
+            coords = [float(coord) for coord in coords]
+            box_left, box_top, box_right, box_bottom = yolobbox2bbox(*coords)
+            box_width = box_right - box_left
+            box_height = box_bottom - box_top
+
+            # Calculate the centre of the bounding box in the original image
+            x = box_right - box_width/2
+            y = box_bottom - box_height/2
+
+            # Calculate the coordinates of the cropped image
+            factor = max(box_width, box_height)
+            factor = max(factor, 320)
+            x1 = int(max(0, min((x - factor // 2), int(img.size[0]) - factor)))
+            y1 = int(max(0, min((y - factor // 2), int(img.size[1]) - factor)))
+            x2 = int(max(factor, min((x + factor // 2), int(img.size[0]))))
+            y2 = int(max(factor, min((y + factor // 2), int(img.size[1]))))
 
         # Crop the image
         img_cropped = img.crop((x1, y1, x2, y2))
@@ -120,14 +156,17 @@ for filename in os.listdir(folder_path):
         new_box_width = new_box_right - new_box_left
         new_box_height = new_box_bottom - new_box_top
 
-        #Convert the coordinates to YOLOv3 format
-        b = (new_box_left, new_box_top, new_box_right, new_box_bottom)
-        bb = bbox2yolobbox((img_resized.size[0], img_resized.size[1]), b)
+        if os.path.exists(txt_path):
+            #Convert the coordinates to YOLOv3 format
+            b = (new_box_left, new_box_top, new_box_right, new_box_bottom)
+            print(b, img_resized.size[0], img_resized.size[1])
+            bb = pbx.convert_bbox(b, from_type="voc", to_type="yolo", image_size=(img_resized.size[0], img_resized.size[1]))
+            #bb = bbox2yolobbox((img_resized.size[0], img_resized.size[1]), b)
 
-        # Write the new coordinates to a new txt file
-        new_txt_path = f"{folder_path}/{filename[:-4]}_320.txt"
-        with open(new_txt_path, "w") as f:
-            f.write(f"0 {round(bb[0], 6)} {round(bb[1], 6)} {round(bb[2], 6)} {round(bb[3], 6)}")
+            # Write the new coordinates to a new txt file
+            new_txt_path = f"{out_folder_path_lbl}/{filename[:-4]}_320.txt"
+            with open(new_txt_path, "w") as f:
+                f.write(f"0 {round(bb[0], 6)} {round(bb[1], 6)} {round(bb[2], 6)} {round(bb[3], 6)}")
 
         img2 = cv2.imread(f"{tmp_folder}/{filename[:-4]}_320.jpg")
 
@@ -144,6 +183,7 @@ for filename in os.listdir(folder_path):
             cv2.destroyAllWindows()
 
         # Move the image and txt file to the original folder
-        shutil.move(destination_path, os.path.join(folder_path, filename))
-        shutil.move(new_img_path, os.path.join(folder_path, f"{filename[:-4]}_320.jpg"))
+        if os.path.exists(txt_path):
+            shutil.move(destination_path, os.path.join(folder_path_img, filename))
+        shutil.move(new_img_path, os.path.join(out_folder_path_img, f"{filename[:-4]}_320.jpg"))
         print(f"File {filename} processed successfully")
